# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tunisian E-Commerce API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  productCount: number;
}

export interface Product {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  price: number;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  categoryId?: number | null;
  /** @nullable */
  categoryName?: string | null;
  stock: number;
  featured: boolean;
  createdAt: string;
}

export interface ProductsPage {
  products: Product[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface CategoryStat {
  categoryId: number;
  categoryName: string;
  count: number;
}

export interface ProductStats {
  totalProducts: number;
  totalCategories: number;
  featuredCount: number;
  byCategory: CategoryStat[];
}

export interface LineItem {
  productId: number;
  productName: string;
  quantity: number;
  unitPrice: number;
  totalPrice: number;
}

export type CheckoutInputPaymentMethod =
  (typeof CheckoutInputPaymentMethod)[keyof typeof CheckoutInputPaymentMethod];

export const CheckoutInputPaymentMethod = {
  cash_on_delivery: "cash_on_delivery",
  flouci: "flouci",
} as const;

export interface CheckoutInput {
  customerName: string;
  email: string;
  phone: string;
  address: string;
  governorate: string;
  city: string;
  paymentMethod: CheckoutInputPaymentMethod;
  items: LineItem[];
  /** @nullable */
  notes?: string | null;
}

export interface Order {
  id: number;
  customerName: string;
  email: string;
  phone: string;
  address: string;
  governorate: string;
  city: string;
  paymentMethod: string;
  status: string;
  subtotal: number;
  shippingFee: number;
  total: number;
  items: LineItem[];
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  flouciPaymentId?: string | null;
  createdAt: string;
}

export interface FlouciInitInput {
  orderId: number;
  amount: number;
  successUrl: string;
  failUrl: string;
}

export interface FlouciInitResult {
  paymentId: string;
  redirectUrl: string;
  status: string;
}

export interface FlouciVerifyResult {
  paymentId: string;
  status: string;
  /** @nullable */
  orderId?: number | null;
}

export type ListProductsParams = {
  /**
   * @nullable
   */
  categoryId?: number | null;
  /**
   * @nullable
   */
  search?: string | null;
  /**
   * @nullable
   */
  featured?: string | null;
  /**
   * @nullable
   */
  page?: number | null;
  /**
   * @nullable
   */
  limit?: number | null;
};