# lib/api-zod/src/generated/types/categoryStat.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tunisian E-Commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface CategoryStat {
  categoryId: number;
  categoryName: string;
  count: number;
}