# lib/api-zod/src/generated/types/checkoutInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tunisian E-Commerce API
 * OpenAPI spec version: 0.1.0
 */
import type { CheckoutInputPaymentMethod } from "./checkoutInputPaymentMethod";
import type { LineItem } from "./lineItem";

export interface CheckoutInput {
  customerName: string;
  email: string;
  phone: string;
  address: string;
  governorate: string;
  city: string;
  paymentMethod: CheckoutInputPaymentMethod;
  items: LineItem[];
  /** @nullable */
  notes?: string | null;
}