# lib/api-zod/src/generated/types/flouciInitInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tunisian E-Commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface FlouciInitInput {
  orderId: number;
  amount: number;
  successUrl: string;
  failUrl: string;
}