# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tunisian E-Commerce API
 * OpenAPI spec version: 0.1.0
 */
import type { LineItem } from "./lineItem";

export interface Order {
  id: number;
  customerName: string;
  email: string;
  phone: string;
  address: string;
  governorate: string;
  city: string;
  paymentMethod: string;
  status: string;
  subtotal: number;
  shippingFee: number;
  total: number;
  items: LineItem[];
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  flouciPaymentId?: string | null;
  createdAt: string;
}