# lib/api-zod/src/generated/types/productStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tunisian E-Commerce API
 * OpenAPI spec version: 0.1.0
 */
import type { CategoryStat } from "./categoryStat";

export interface ProductStats {
  totalProducts: number;
  totalCategories: number;
  featuredCount: number;
  byCategory: CategoryStat[];
}