# lib/api-zod/src/generated/types/productsPage.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tunisian E-Commerce API
 * OpenAPI spec version: 0.1.0
 */
import type { Product } from "./product";

export interface ProductsPage {
  products: Product[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}